refactor(utils): make utils its own package
